refactor: ordering files in different folders
