--- LivlReviews.Domain.Test/Test1.cs
@@ -1,6 +1,0 @@
-namespace LivlReviews.Domain.Test;
-
-public class Test1
-{
-    // TODO : implement logic and test it
-}
