+++ LivlReviews.Domain/Domain_interfaces_input/IInvitationConfirmator.cs
