+++ LivlReviews.Domain/Domain_interfaces_input/IInvitationSender.cs
