+++ LivlReviews.Domain/Domain_interfaces_output/IInvitationDelivery.cs
