+++ LivlReviews.Domain/Domain_interfaces_output/IInvitationTokenInventory.cs
