+++ LivlReviews.Domain/Domain_interfaces_output/IUserInventory.cs
